refactor(agentic-ai): move ResponseConfiguration to an interface and flag to include agent context to JW specific implementation
